add : 포인트로그,페이먼트 로그 페이징 및 테스트 추가,  어카운트 테스트 추가

+++ market/src/main/java/com/charly/market/point/model/dto/CreatePointLogRequest.java
@@ -4,11 +4,12 @@ package com.charly.market.point.model.dto;
 import jakarta.validation.constraints.NotBlank;
 
 public record CreatePointLogRequest(
-        @NotBlank String type,
-        @NotBlank Long tradeAmount,
-        @NotBlank String explanation,
-        @NotBlank Long auctionBidId,
-        @NotBlank Long userId,
-        @NotBlank int pointAmount
+    @NotBlank String type,
+    @NotBlank Long tradeAmount,
+    @NotBlank String explanation,
+    @NotBlank Long auctionBidId,
+    @NotBlank Long userId,
+    @NotBlank Long pointAmount
 ) {
+
 }
